pyproject.toml
[project]
name = "my-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.5.6",
    "autogen-ext[openai]>=0.5.6",
    "autogenstudio>=0.1.5",
    "python-decouple>=3.8",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.scripts]
my-ai-agents = "my_ai_agents:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
]
